refactor: rename types folder to typings
