+++ src/typings/obsidian-ex.d.ts
